rename: galleryContent 파일 이동으로 인한 코드 수정
